Fix incompatibilities with newer changes from kgb.
kgb's signature compatibility is getting more strict, and we have a few
functions passed in that don't match the signatures for the original
functions. We also have a direct assignment for a variable to reset the
calls on a spy, which wasn't correct.

This change fixes both by using the proper variable names for the
function being overridden and calling `reset_calls()` to reset the call
state.

Testing Done:
Unit tests pass with the new kgb and old kgb.

Reviewed at https://reviews.reviewboard.org/r/9521/