Fix showing the current avatar in the File Upload config form.
The File Upload avatar configuration form was checking the wrong
configuration variable to see if an avatar was set. That resulted in the
avatar always appearing as "missing." This change simply switches it to
the correct variable so that the current avatar is shown.

Testing Done:
Uploaded an avatar. Saw that the avatar persisted in the config form
after page reload.

Reviewed at https://reviews.reviewboard.org/r/9485/