Register global defaults for the avatar configuration.
The avatars registry was registering defaults on the current site
configuration when first populated, but this meant that code couldn't
rely on these defaults until accessing the registry. Now, global
defaults are registered instead, allowing immediate access so long as
the module has been imported.

In the future, we'd want to move this to `AppConfig.ready()`, but for
now this will suffice.

Testing Done:
Unit tests pass.

Verified the default values manually.

Reviewed at https://reviews.reviewboard.org/r/9590/