Deprecate ConfigPageForm.profile.
When config pages were moved from Review Board to Djblets, the
`ConfigPageForm.profile` attribute remained. This was never used in
Djblets, and with `User.get_profile()` no longer being a thing in Django
itself, it doesn't make sense to have this here.

This is now deprecated. Instead of setting it automatically in the
constructor, it's now a dynamic property that warns when accessed.
Subclasses are advised to fetch their own concept of a profile
themselves.

A unit test was added to test the warning.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9618/