Switch to calling print as a function in Djblets.
We have a few cases where we print content to the console using the
`print` statement (all for packaging). This change switches these to
`print()` calls, as needed on Python 3.

Testing Done:
Tested that packaging worked on Python 2.7. We're not ready to test
this on Python 3 yet (but `setup.py` did not immediately blow up).

Reviewed at https://reviews.reviewboard.org/r/9660/