Switch to using mysqlclient instead of MySQL-python as a dev dependency.
`MySQL-python` hasn't really been maintained in a while, is not
compatible with Python 3, and isn't really getting a whole lot of
updates. The successor to this package, which is fully API-compatible,
has wheels, and is compatible with Python 2.7 and 3, is `mysqlclient`.
This switches our development dependency over to that.

Testing Done:
Successfully installed via pip and easy_install on Python 2.7, 3.4,
3.5, and 3.6. MySQL-specific unit tests passed with this dependency
(as did Django Evolution tests for MySQL, which is a better set of
tests in this case).

Reviewed at https://reviews.reviewboard.org/r/9671/