Switch to the modern format for octals.
Python 2.6 introduced a new format for octals, which is in the form of
`0o755` instead of `0755`. The new format was optional through Python
2.x, but is mandatory in Python 3.

This change simply switches all of our octals in the codebase to the
modern format.

Testing Done:
Unit tests pass on Python 2.7. We're not in a position yet to run tests
on Python 3.

Reviewed at https://reviews.reviewboard.org/r/9672/