Fix function signatures for logging functions on Python 3.
We attempt to work around an issue with some older versions of the
`logging` functions that didn't properly work with keyword arguments.
We used `inspect.getargspec()` to check for this, but this function is
deprecated and doesn't take into account keyword-only arguments, and
causes warnings on Python 3.

This change switches this over to use `inspect.getfullargspec()` when
available.

Testing Done:
Ran tests (along with other upcoming changes) on Python 2.7, 3.4, 3.5,
and 3.6. I no longer hit any warnings with the logging calls.

Reviewed at https://reviews.reviewboard.org/r/9726/