Update legacy force_unicode imports to use force_text instead.
Django provided an old `django.utils.encoding.force_unicode` function,
which was only importable on Python 2.7. The "new" name (compatible with
Django 1.6) for this is `force_text`, which is importable on all Python
versions. We had a few modules importing the old name, which caused them
to fail to load and in turn broke the Django apps, preventing unit tests
from running.

This change updates those imports for the new names, allowing the test
suite to at least attempt to execute.

Testing Done:
The test suite was able to run on Python 2.7, 3.4, 3.5, and 3.6 without
immediately and entirely failing. Along with all upcoming changes, the
tests involving these imports pass.

Reviewed at https://reviews.reviewboard.org/r/9727/