Fix some simple cases of using the wrong string type.
We've tried to be future-proof with our string types, using Unicode
strings for most things and byte strings where we thought we needed
them. These weren't always correct, though. There are strings that
needed to be the native string type for the Python version in some
cases, and other places where we had to go back to literal byte strings
or to encode to those.

This change takes care of most of the simpler cases. There are still
other places in the codebase where the wrong string type is being used,
but they're part of bigger changes that will be addressed separately.

Testing Done:
Unit tests pass on Python 2.7, 3.4, 3.5, and 3.6 (when used with other
upcoming changes).

Reviewed at https://reviews.reviewboard.org/r/9728/