Remove the nesting of classes in RelationCounterField.
This is the first in a series of changes to simplify the complexity in
`RelationCounterField`, disentangling the responsibilities between
classes and reducing chances of deadlocks or other issues.

In this change, the nested classes (`InstanceState` and
`RelationTracker`) have been moved out from `RelationTrackerField` into
the top level of the new module for the field. This just helps keep
things a bit more organized for the next few steps and reduces the
amount of churn that has to be reviewed.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9755/