Update some calls and missing functions needed in Python 3.
There's a handful of legacy calls, or legacy forms of calls, that needed
to be updated to be compatible with Python 2.x and 3.x. These have to do
with module reloading and importing and generator iteration. There's
also a missing `__hash__` needed for an object in a unit test that
otherwise prevented its storage in a dictionary.

Testing Done:
Unit tests pass on Python 2.7, 3.4, 3.5, and 3.6.

Reviewed at https://reviews.reviewboard.org/r/9736/