Disable looking for and executing doctests.
Our test runner code was attempting to look for code in documentation to
run when running the unit test suite. This isn't necessary, and actually
causes problems with Python 3 (it attempts to look up something in
Pipeline's configuration and then exits with an exception). We don't use
doctests, and don't need to spend time looking for them, so turn off
this part of the test run.

Testing Done:
Unit tests run on Python 2.7, 3.4, 3.5, and 3.6.

Reviewed at https://reviews.reviewboard.org/r/9737/