Convert data in responses and resource streams to the right string type.
`HttpResponse` returns data as byte strings, and in Python 2.x we could
generally just check against those without problems, but on Python 3.x
the different string types are less compatible. We now convert the
responses to Unicode strings using Django's `force_text()` (which
intelligently performs a decode when necessary).

We also convert the result of `pkg_resources.resource_stream` for the
DMARC code, for use in `PublicSuffixList`. Rather than a straight text
conversion, we need to make use of `codecs.getreader()`, which will
correctly convert the data in the stream to the right format.

Testing Done:
Unit tests pass on Python 2.7, 3.4, 3.5, and 3.6.

Reviewed at https://reviews.reviewboard.org/r/9735/