Fix JavaScript on extensions config page.
The way we were instantiating the collection for extensions on the
config page (passing the URL in as an option and expecting it to get
attached) wasn't officially supported, and doesn't work with the latest
backbone.js. This change moves it to assign the URL after creating the
collection.

Testing done:
Loaded the extensions list and saw that there were no more JavaScript
errors.

Reviewed at https://reviews.reviewboard.org/r/9850/