Don't render form field labels when the label is an empty string.
Django forms allow for the disabling of `<label>` tags if the label is
an empty string, but our counterpart (`{% label_tag %}`) didn't support
this. We now look to see if a field's label is an empty string and, if
so, we simply don't render a label for the field.

Testing Done:
Unit tests passed.

Tested along with the upcoming consent change.

Reviewed at https://reviews.reviewboard.org/r/9923/